Classe Employe 2, change Integer to Long

+++ src/main/java/com/ipiecoles/java/java230/repository/EmployeRepository.java
@@ -3,5 +3,5 @@ package com.ipiecoles.java.java230.repository;
 import com.ipiecoles.java.java230.model.Employe;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeRepository extends CrudRepository<Employe,Integer> {
+public interface EmployeRepository extends CrudRepository<Employe,Long> {
 }
